+++ packages/interactive_media_ads/test/android/ads_loader_test.mocks.dart
@@ -527,13 +527,6 @@ class MockAdsLoadedListener extends _i1.Mock implements _i2.AdsLoadedListener {
 /// See the documentation for Mockito's code generation for more information.
 class MockAdsManager extends _i1.Mock implements _i2.AdsManager {
   @override
-  List<double?> get adCuePoints => (super.noSuchMethod(
-        Invocation.getter(#adCuePoints),
-        returnValue: <double?>[],
-        returnValueForMissingStub: <double?>[],
-      ) as List<double?>);
-
-  @override
   _i2.PigeonInstanceManager get pigeon_instanceManager => (super.noSuchMethod(
         Invocation.getter(#pigeon_instanceManager),
         returnValue: _FakePigeonInstanceManager_0(
@@ -575,6 +568,16 @@ class MockAdsManager extends _i1.Mock implements _i2.AdsManager {
         returnValue: _i6.Future<void>.value(),
         returnValueForMissingStub: _i6.Future<void>.value(),
       ) as _i6.Future<void>);
+
+  @override
+  _i6.Future<List<double?>> getAdCuePoints() => (super.noSuchMethod(
+        Invocation.method(
+          #getAdCuePoints,
+          [],
+        ),
+        returnValue: _i6.Future<List<double?>>.value(<double?>[]),
+        returnValueForMissingStub: _i6.Future<List<double?>>.value(<double?>[]),
+      ) as _i6.Future<List<double?>>);
 
   @override
   _i6.Future<void> resume() => (super.noSuchMethod(